board: rzv2l-dev: Update change vmmc and vqmmc for eSD mode
Currently, default SDHI0 mode is eMMC from kernel devicetree
for RZ/V2L Smarc.
So, when eSD boot mode is used, uSD can not be used in kernel space.
To support uSD for SDHI0 without changing kernel devicetree setting,
fdt is used to change the vmmc and vccq regulators to gpio regulators
which are used for uSD.

Signed-off-by: Au Dang <au.dang.jx@renesas.com>
Signed-off-by: Linh Hua <linh.hua.jc@renesas.com>